requirements.txt
python-telegram-bot==13.0
PyYAML
fpdf